Update: Add credits. Fixed legend strings. Tie level selector to cheats. Add platform errors for missing/invalid files. Add add savegames to appdata. Disable screenshots (broken) Fix crash on empty sample buffers. Warn about encrypted levels. Add show_logo_in_title flag Change options casing.

+++ TOMB4/game/trng/trng_script_parser.h
@@ -385,8 +385,8 @@ extern char *NGGetString(short string_id);
 extern char *NGGetPluginString(short plugin_id);
 extern int NGGetT4PluginID(short plugin_id);
 
-extern void NGScriptInit(char* gfScriptFile, unsigned int offset, unsigned int len);
+extern void NGScriptInit(char* gfScriptFile, unsigned int offset, size_t len);
 extern void NGScriptCleanup();
 extern void NGLoadTablesForLevel(unsigned int level);
-extern void NGReadNGGameflowInfo(char* gfScriptFile, unsigned int offset, unsigned int len);
-extern void NGReadNGExtraStrings(char* gfLanguageFile, unsigned int offset, unsigned int len);
+extern void NGReadNGGameflowInfo(char* gfScriptFile, unsigned int offset, size_t len);
+extern void NGReadNGExtraStrings(char* gfLanguageFile, unsigned int offset, size_t len);
